demo AuthenticationRouter
