+++ package.json
@@ -19,6 +19,10 @@
     "test:watch": "nwb test-react --server"
   },
   "dependencies": {
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
+    "oidc-client": "^1.10.1",
+    "jwt-decode": "^2.2.0",
     "react-intl": "^3.11.0",
     "react-request-fullscreen": "^1.1.2",
     "prop-types": "15.7.2",
